feat: added check for auth zipfile download and safe_unzip function
The check for the auth zipfile is unique to HHIT and needs to line up with the config.json zipfile url. The safe_unzip function will ensure that the files being unzipped are not being unzipped to other directories but only into the current working directory.

+++ src/manage.py
@@ -193,17 +193,33 @@ class Manage:
             from io import BytesIO
             from zipfile import ZipFile
 
-            self.handleProxy()
-            r = requests.get(self.tool_zipfile_url)
+            def safe_unzip(zip_file, extract_path='.'):
+                with ZipFile(zip_file, 'r') as zf:
+                    for member in zf.infolist():
+                        file_path = os.path.realpath(os.path.join(extract_path, member.filename))
+                        if file_path.startswith(os.path.realpath(extract_path)):
+                            zf.extract(member, extract_path)
 
-            z = ZipFile(BytesIO(r.content))
-            z.extractall()
-            fromDirectory = z.namelist()[0]
-            toDirectory = './'
-            copy_tree(fromDirectory, toDirectory)
-            rmtree(fromDirectory)
-            self.messageBox(
-                0, u'The tool was successfully updated! I hope that was quick enough for you. The update will take effect when the tool is reopened.', u"HazPy", 0x1000 | 0x4)
+            self.handleProxy()
+            
+            if self.tool_zipfile_url in ["https://github.com/nhrap-dev/HHIT/archive/dev.zip",
+                                         "https://github.com/nhrap-hazus/HHIT/archive/test.zip",
+                                         "https://github.com/nhrap-hazus/HHIT/archive/main.zip"]:
+                r = requests.get(self.tool_zipfile_url)
+                z = ZipFile(BytesIO(r.content))
+                #z.extractall()
+                print('safe unzip start')
+                safe_unzip(z)
+                print('safe unzip stop')
+                fromDirectory = z.namelist()[0]
+                toDirectory = './'
+                copy_tree(fromDirectory, toDirectory)
+                rmtree(fromDirectory)
+                self.messageBox(
+                    0, u'The tool was successfully updated! I hope that was quick enough for you. The update will take effect when the tool is reopened.', u"HazPy", 0x1000 | 0x4)
+            else:
+                self.messageBox(
+                    0, u'The tool update failed: bad zipefile config. If this error persists, contact hazus-support@riskmapcds.com for assistance.', u"HazPy", 0x1000 | 0x4)
         except:
             self.messageBox(
                 0, u'The tool update failed. If this error persists, contact hazus-support@riskmapcds.com for assistance.', u"HazPy", 0x1000 | 0x4)
